Refactor logger tests
